+++ lib/http_client.dart
@@ -1,0 +1,17 @@
+/*
+* http_client.dart
+* RestClient.
+* Dashkevich Andrey, 13 March 2024
+*/
+
+/// Rest client.
+///
+/// A simple lightweight wrapper around [http](https://pub.dev/packages/http).
+///
+/// Package has simple utils library.
+/// See [utils]({@macro utils}).
+library;
+
+export 'package:http/http.dart';
+export 'src/http_methods.dart';
+export 'src/client.dart';