+++ lib/utils.dart
@@ -2,7 +2,7 @@
 /// Use simple utils from json serialization [JsonParser]
 /// or parse jwt token [JWT],[JwtDecoder].
 /// {@endtemplate}
-library utils;
+library;
 
 export 'src/utils/json_parser/json_parser.dart';
 export 'src/utils/jwt/jwt.dart';